+++ src/main/java/baseball/Application.java
@@ -1,5 +1,7 @@
 package baseball;
 
+import baseball.controller.BaseBallGame;
+
 public class Application {
     public static void main(String[] args) {
         BaseBallGame baseBallGame = new BaseBallGame();
